Add files via upload
We developed a new mean regression model for the zero-truncated Poisson distribution and its generalization to k-th truncated Poisson distribution. The provided R programs are used to calculate its bis, MSE, empirical power, significant level in simulation; to calculate the MLEs of parameters for the ZTP regression model in real data analysis; to compare the two existing ZTP regression models and the provided mean regression model.